requirements.txt
tensorflow-macos>=2.8.0
tensorflow-metal>=0.5.0
numpy>=1.19.2
scipy>=1.7.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
h5py>=3.1.0
PyWavelets>=1.2.0
keras-tuner>=1.1.0
tqdm>=4.62.3